refactor: restructure folders to be more organized and accessible
