+++ lib/app/core/errors/exceptions.dart
