+++ lib/app/core/errors/failure.dart
