+++ lib/app/core/utilities/helpers/widget_helper.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:rock_n_roll_forecast/app/core/theme/text_theme.dart';
+import 'package:rock_n_roll_forecast/app/presentation/theme/text_theme.dart';
 
-import '../../theme/app_colors.dart';
+import '../../../presentation/theme/app_colors.dart';
 
 class WidgetHelper {
   static Widget error(String text) {
