+++ lib/app/data/datasources/local_datasource/local_datasource_impl.dart
@@ -2,13 +2,13 @@ import 'dart:async';
 
 import 'package:flutter/material.dart';
 import 'package:hive/hive.dart';
-import 'package:rock_n_roll_forecast/app/core/utilities/errors/exceptions.dart';
+import 'package:rock_n_roll_forecast/app/core/errors/exceptions.dart';
 import 'package:rock_n_roll_forecast/app/domain/entities/forecast_entity.dart';
 import 'package:rock_n_roll_forecast/app/domain/entities/weather_entity.dart';
 
+import '../../../core/errors/failure.dart';
 import '../../../core/utilities/adapters/local_storage_adapter/local_storage_adapter.dart';
 import '../../../core/utilities/constants.dart';
-import '../../../core/utilities/errors/failure.dart';
 import 'local_datasource.dart';
 
 class WeatherLocalDatasourceImpl implements WeatherLocalDatasource {
