+++ lib/app/domain/repositories/local_repository.dart
@@ -2,7 +2,7 @@ import 'package:dartz/dartz.dart';
 import 'package:rock_n_roll_forecast/app/domain/entities/forecast_entity.dart';
 import 'package:rock_n_roll_forecast/app/domain/entities/weather_entity.dart';
 
-import '../../core/utilities/errors/failure.dart';
+import '../../core/errors/failure.dart';
 
 abstract class WeatherLocalRepository {
   Future<Either<Failure, void>> cacheWeather(
