+++ lib/app/domain/usecases/local_usecases/cache_forecast_usecase.dart
@@ -1,7 +1,7 @@
 import 'package:dartz/dartz.dart';
 import 'package:rock_n_roll_forecast/app/domain/repositories/local_repository.dart';
 
-import '../../../core/utilities/errors/failure.dart';
+import '../../../core/errors/failure.dart';
 import '../../entities/forecast_entity.dart';
 
 abstract class CacheForecastUseCase {
