+++ lib/app/domain/usecases/local_usecases/cache_weather_usecase.dart
@@ -2,7 +2,7 @@ import 'package:dartz/dartz.dart';
 import 'package:rock_n_roll_forecast/app/domain/entities/weather_entity.dart';
 import 'package:rock_n_roll_forecast/app/domain/repositories/local_repository.dart';
 
-import '../../../core/utilities/errors/failure.dart';
+import '../../../core/errors/failure.dart';
 
 abstract class CacheWeatherUsecase {
   Future<Either<Failure, void>> execute(
