+++ lib/app/domain/usecases/local_usecases/offline_weather_usecase.dart
@@ -1,6 +1,6 @@
 import 'package:dartz/dartz.dart';
 
-import '../../../core/utilities/errors/failure.dart';
+import '../../../core/errors/failure.dart';
 import '../../entities/weather_entity.dart';
 import '../../repositories/local_repository.dart';
