+++ lib/app/presentation/routes/routes_builder.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:rock_n_roll_forecast/app/core/routes/routes.dart';
+import 'package:rock_n_roll_forecast/app/presentation/routes/routes.dart';
 import 'package:rock_n_roll_forecast/app/presentation/screens/concert_info_screen.dart';
 
-import '../../presentation/screens/concerts_screen.dart';
+import '../screens/concerts_screen.dart';
 
 Route<dynamic> routesController(RouteSettings settings) {
   switch (settings.name) {
