+++ lib/app/presentation/theme/app_theme.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:rock_n_roll_forecast/app/core/theme/input_theme.dart';
-import 'package:rock_n_roll_forecast/app/core/theme/text_theme.dart';
+import 'package:rock_n_roll_forecast/app/presentation/theme/input_theme.dart';
+import 'package:rock_n_roll_forecast/app/presentation/theme/text_theme.dart';
 
-import '../utilities/constants.dart';
+import '../../core/utilities/constants.dart';
 import 'app_color_scheme.dart';
 
 class AppTheme {
