+++ lib/app/presentation/widgets/concert_title.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 
-import '../../core/theme/app_colors.dart';
-import '../../core/theme/text_theme.dart';
 import '../../core/utilities/helpers/responsive_helper.dart';
+import '../theme/app_colors.dart';
+import '../theme/text_theme.dart';
 
 class ConcertTitle extends StatelessWidget {
   const ConcertTitle({
