+++ lib/app/presentation/widgets/weather_forecast_loader.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:shimmer/shimmer.dart';
 
-import '../../core/theme/app_colors.dart';
 import '../../core/utilities/helpers/responsive_helper.dart';
+import '../theme/app_colors.dart';
 
 class WeatherForecastLoader extends StatelessWidget {
   const WeatherForecastLoader({super.key});
