+++ test/app/core/routes/routes_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:rock_n_roll_forecast/app/core/routes/routes.dart';
+import 'package:rock_n_roll_forecast/app/presentation/routes/routes.dart';
 
 void main() {
   // test group: Routes
