+++ test/app/core/utilities/errors/failure_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:rock_n_roll_forecast/app/core/utilities/errors/failure.dart';
+import 'package:rock_n_roll_forecast/app/core/errors/failure.dart';
 
 void main() {
   group('Failure', () {
