+++ test/app/core/utilities/helpers/test_helper.dart
@@ -4,9 +4,9 @@ import 'package:flutter/services.dart';
 import 'package:hive/hive.dart';
 import 'package:http/http.dart';
 import 'package:mockito/annotations.dart';
+import 'package:rock_n_roll_forecast/app/core/errors/exceptions.dart';
 import 'package:rock_n_roll_forecast/app/core/utilities/adapters/connectivity_adapter/connectivity_adapter.dart';
 import 'package:rock_n_roll_forecast/app/core/utilities/adapters/local_storage_adapter/local_storage_adapter.dart';
-import 'package:rock_n_roll_forecast/app/core/utilities/errors/exceptions.dart';
 import 'package:rock_n_roll_forecast/app/core/utilities/helpers/location_helper.dart';
 import 'package:rock_n_roll_forecast/app/core/utilities/helpers/misc_helper.dart';
 import 'package:rock_n_roll_forecast/app/data/datasources/local_datasource/local_datasource.dart';
