+++ test/app/data/datasources/local_datasource_test.dart
@@ -1,9 +1,9 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:hive/hive.dart';
 import 'package:mockito/mockito.dart';
+import 'package:rock_n_roll_forecast/app/core/errors/exceptions.dart';
+import 'package:rock_n_roll_forecast/app/core/errors/failure.dart';
 import 'package:rock_n_roll_forecast/app/core/utilities/constants.dart';
-import 'package:rock_n_roll_forecast/app/core/utilities/errors/exceptions.dart';
-import 'package:rock_n_roll_forecast/app/core/utilities/errors/failure.dart';
 import 'package:rock_n_roll_forecast/app/data/datasources/local_datasource/local_datasource_impl.dart';
 
 import '../../core/utilities/helpers/test_helper.mocks.dart';
